backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.6,<5.0
dj-database-url>=2.2,<3.0
Pillow>=11.0,<12.0
psycopg2-binary>=2.9,<3.0
